Added option to start new

+++ dbtools/src/main/scala/DBMigration.scala
@@ -10,7 +10,7 @@ import io.iohk.iodb.ByteArrayWrapper
 
 import scala.collection.mutable.ArrayBuffer
 import scala.reflect.io.Path
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 import java.io.File
 import java.io.PrintWriter
 
@@ -53,41 +53,53 @@ object DBMigration extends Logging {
     log.info(s"Found all blockIds in $duration seconds")
   }
 
-  private def migrate(oldSettings: AppSettings, newSettings: AppSettings): Unit = {
+  private def migrate(oldSettings: AppSettings, newSettings: AppSettings, startNew: Boolean): Unit = {
 
-    /* If new-data data exists, remove before creating new-data */
-    val path: Path = Path(".bifrost/new-data")
-    Try(path.deleteRecursively())
+    if (startNew) {
+      /* If new-data data exists, remove before creating new-data */
+      val path: Path = Path(".bifrost/new-data")
+      Try(path.deleteRecursively())
+    }
 
     val idsFile = Source.fromFile(".bifrost/blockIds/bids.txt")
     val oldHistory: History = History.readOrGenerate(oldSettings)
     val newHistory: History = History.readOrGenerate(newSettings)
     val newState: State = State.readOrGenerate(newSettings, callFromGenesis = true, newHistory)
-    var height: Long = 1
+    var height: Long = newHistory.height
     val start = System.nanoTime()
 
     var parentBlockId: ModifierId = ModifierId(History.GenesisParentId)
 
-    idsFile.getLines.foreach{ line =>
+    idsFile.getLines.drop(height.toInt).foreach{ line =>
       val bid: Array[Byte] = Base58.decode(line).get
+
       val currentBlock: Block = oldHistory.storage.storage.get(ByteArrayWrapper(bid)).map { bw =>
         val bytes = bw.data
         BlockSerializer.decode(bytes.tail).get
       }.get.copy(parentId = parentBlockId)
+
       val currentDifficulty: Long = oldHistory.storage.storage.get(ByteArrayWrapper(bid)).map { bw =>
         val bytes = bw.data
         Longs.fromByteArray(bytes)
       }.get
-      log.debug(s"----------------------------------------------------------------------------------------------------------------------Height:$height")
+
       height = height + 1
-      newHistory.storage.update(currentBlock, currentDifficulty, isBest = true)
-      newState.applyModifierOnState(currentBlock)
+
+      Try {
+        newHistory.storage.update(currentBlock, currentDifficulty, isBest = true)
+        newState.applyModifierOnState(currentBlock)
+      } match {
+        case Success(_) => log.debug(s"----------------------------------------------------------------------------------------------------------------------Height:$height")
+        case Failure(_) => log.warn(s"Failed to append and apply block $height !!!!!!!")
+      }
+
       parentBlockId = currentBlock.id
     }
 
     idsFile.close()
     val duration = (System.nanoTime() - start) / 1e9d
     height -= 1
+
     log.info(s"Migrated $height blocks in $duration seconds")
   }
 
@@ -100,6 +112,6 @@ object DBMigration extends Logging {
     val newSettings: AppSettings = AppSettings.read(StartupOpts(Some(newSettingsFilename), None))
 
 //    getIds(oldSettings)
-    migrate(oldSettings, newSettings)
+    migrate(oldSettings, newSettings, false)
   }
 }
